src/requirements.txt
certifi==2019.3.9
chardet==3.0.4
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.8
monotonic==1.5
python-engineio==3.4.4
python-pytuntap==1.0.5
python-socketio==4.0.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
